requirements.txt
-e git+https://github.com/redhat-cip/pysflib.git#egg=pysflib
-e git+https://github.com/openstack-infra/python-storyboardclient.git#egg=storyboardclient
pecan==1.0.3
WebOb==1.5.0a1
gerritlib==0.4.0
python-ldap==2.4.20
requests==2.7.0
passlib==1.6.5
basicauth==0.3
sqlalchemy==1.0.8
urllib3==1.11
paramiko
pycrypto==2.6.1
htpasswd==2.3
pygerrit # see dependency in pysflib
python-redmine==1.2.0
sphinx==1.3.1
PyYAML==3.11
stevedore==1.7.0
MySQL-python==1.2.5
six==1.9.0
deepdiff==2.1.2
